requirements.txt
grpcio-tools
mypy-protobuf

git+https://github.com/Ultimaker/CuraEngine_grpc_defintions@main#egg=CuraEngineGRPC


# Example simplification requirement
svgpathtools